pyproject.toml
[tool.poetry]
name = "paido-backend"
version = "0.1.0"
description = ""
authors = ["hugo rodriguez <hrodrime@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.*"
fastapi = {extras = ["standard"], version = "^0.115.6"}
pydantic = {extras = ["email"], version = "^2.10.4"}
sqlalchemy = "^2.0.36"
pydantic-settings = "^2.7.1"
alembic = "^1.14.0"
pwdlib = {extras = ["argon2"], version = "^0.2.1"}
python-multipart = "^0.0.20"
pyjwt = "^2.10.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.8.4"
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
taskipy = "^1.14.1"

[tool.taskipy.tasks]
run = 'fastapi dev paido_backend/app.py'
pre_test = 'task lint'
test = 'pytest --cov=paido_backend -vv'
post_test = 'coverage html'
lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . --fix && ruff format .'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.ruff]
line-length = 90
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I','F','E','W','PL','PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
